initialize mixer with MP3
This works as long as SDL2_mixer is built with MP3 support.
The homebrew version on my mac was lacking that, but after building it
from source it worked like a charm.

Making this optional, so the game doesn't fail to start just
because music is missing, and because others might have SDL_mixer
without mp3 support.